Breakout `equalizeHistogram()`
