+++ src/util/equalize-histogram.js
@@ -1,0 +1,301 @@
+/**
+ * Real-time object detector based on the Viola Jones Framework.
+ * Compatible to OpenCV Haar Cascade Classifiers (stump based only).
+ *
+ * Copyright (c) 2017, Martin Tschirsich, Alex Lemanski
+ */
+
+/**
+ * Equalizes the histogram of an unsigned 1-channel image with integer
+ * values in [0, 255]. Corresponds to the equalizeHist OpenCV function.
+ *
+ * @param {Array}  src   1-channel integer source image
+ * @param {Number} step  Sampling stepsize, increase for performance
+ * @param {Array}  [dst] 1-channel destination image
+ *
+ * @return {Array} 1-channel destination image
+ */
+export default function equalizeHistogram(src, step, dst) {
+  var srcLength = src.length;
+  if (!dst) dst = src;
+  if (!step) step = 5;
+
+  // Compute histogram and histogram sum:
+  var hist = [
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0,
+    0
+  ];
+
+  for (var i = 0; i < srcLength; i += step) {
+    ++hist[src[i]];
+  }
+
+  // Compute integral histogram:
+  var norm = 255 * step / srcLength,
+    prev = 0;
+  for (var i = 0; i < 256; ++i) {
+    var h = hist[i];
+    prev = h += prev;
+    hist[i] = h * norm; // For non-integer src: ~~(h * norm + 0.5);
+  }
+
+  // Equalize image:
+  for (var i = 0; i < srcLength; ++i) {
+    dst[i] = hist[src[i]];
+  }
+  return dst;
+}
